feat(mmbot): add more important changement
- parsing float api inpoint into string
- start http rest call getprice
- add more unit tests

+++ src/utils/mmbot_strong_types.hpp
@@ -23,7 +23,7 @@
 namespace antara
 {
     static constexpr const int g_factor = 100000000;
-    static constexpr const double g_rounding = 0.000000004999;
+    static constexpr const long double g_rounding = 0.000000004999;
     using st_endpoint = st::type<
             std::string,
             struct endpoint_tag,
